refactor: split themes api into focused router modules
