+++ .beads/issues.jsonl
@@ -454,7 +454,7 @@
 {"id":"StockScreenClaude-xz9","title":"Implement instant snapshot-backed full-universe scans","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-03-19T01:00:15.6989+08:00","created_by":"David Ten","updated_at":"2026-03-19T08:33:19.834264+08:00","closed_at":"2026-03-19T08:33:19.834264+08:00","close_reason":"Implemented instant snapshot-backed full-universe scan path"}
 {"id":"StockScreenClaude-y5c","title":"Fix rate limit constant in warm_all_caches()","description":"cache_manager.py line 221 uses scan_rate_limit (1.0s) instead of yfinance_batch_rate_limit_interval (5.0s)","status":"closed","priority":1,"issue_type":"bug","owner":"davidten7@gmail.com","created_at":"2026-02-19T08:11:44.99013+08:00","created_by":"David Ten","updated_at":"2026-02-19T08:12:09.721103+08:00","closed_at":"2026-02-19T08:12:09.721103+08:00","close_reason":"Fixed scan_rate_limit -\u003e yfinance_batch_rate_limit_interval in warm_all_caches()"}
 {"id":"StockScreenClaude-y8bv","title":"Fix tranche-1 review findings","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-04-03T08:43:19.527587+08:00","created_by":"David Ten","updated_at":"2026-04-03T08:45:56.279881+08:00","closed_at":"2026-04-03T08:45:56.279881+08:00","close_reason":"Completed"}
-{"id":"StockScreenClaude-y9af","title":"Decompose theme platform routers and services","description":"Break backend/app/api/v1/themes.py into focused routers and move theme orchestration out of the route layer. Create a pragmatic theme-platform slice with small deterministic services, query/repository ownership for SQL, and no corruption-recovery DDL inside API modules.","status":"open","priority":1,"issue_type":"epic","owner":"davidten7@gmail.com","created_at":"2026-04-08T11:45:19.125111+08:00","created_by":"David Ten","updated_at":"2026-04-08T13:22:08.732358+08:00"}
+{"id":"StockScreenClaude-y9af","title":"Decompose theme platform routers and services","description":"Break backend/app/api/v1/themes.py into focused routers and move theme orchestration out of the route layer. Create a pragmatic theme-platform slice with small deterministic services, query/repository ownership for SQL, and no corruption-recovery DDL inside API modules.","status":"in_progress","priority":1,"issue_type":"epic","owner":"davidten7@gmail.com","created_at":"2026-04-08T11:45:19.125111+08:00","created_by":"David Ten","updated_at":"2026-04-08T17:00:47.866448+08:00"}
 {"id":"StockScreenClaude-y9bu","title":"Disable Redis for static-site build workflows","status":"open","priority":1,"issue_type":"bug","owner":"davidten7@gmail.com","created_at":"2026-04-02T10:24:41.590671+08:00","created_by":"David Ten","updated_at":"2026-04-02T10:24:41.590671+08:00"}
 {"id":"StockScreenClaude-yfz","title":"T17: Update existing scan tests for universe changes","status":"closed","priority":2,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-02-16T14:49:45.931233+08:00","created_by":"David Ten","updated_at":"2026-02-16T14:58:24.266684+08:00","closed_at":"2026-02-16T14:58:24.266684+08:00","close_reason":"Implemented in universe unification session","dependencies":[{"issue_id":"StockScreenClaude-yfz","depends_on_id":"StockScreenClaude-bkd","type":"blocks","created_at":"2026-02-16T14:50:51.393973+08:00","created_by":"David Ten"}]}
 {"id":"StockScreenClaude-yj6","title":"Validate Windows desktop installer on a clean machine","description":"Compile backend/desktop/windows_installer.iss, install the generated Setup.exe on a clean Windows machine, verify first-run bootstrap, scan execution, uninstall/reinstall behavior, and data persistence under LOCALAPPDATA.","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-03-17T09:18:04.017064+08:00","created_by":"David Ten","updated_at":"2026-04-08T12:03:27.912565+08:00","closed_at":"2026-04-08T12:03:27.912565+08:00","close_reason":"Superseded by server-only roadmap; desktop installer validation is no longer part of the supported mainline deployment path."}
