+++ backend/app/api/v1/themes.py
@@ -1,123 +1,17 @@
-"""
-API endpoints for Theme Discovery.
+"""Aggregator router for themes feature surfaces.
 
-Provides access to:
-- Theme rankings and metrics
-- Emerging theme discovery
-- Content source management
-- Theme validation and correlation analysis
-- Alerts
+This module also re-exports legacy callables used by existing tests and
+internal imports while the router decomposition is rolling out.
 """
-import csv
-import io
+
+from __future__ import annotations
+
 import logging
-from collections import Counter, defaultdict
-from pydantic import ValidationError
-from fastapi import APIRouter, Depends, HTTPException, Query, BackgroundTasks, Header, Request
-from fastapi.responses import StreamingResponse
-from sqlalchemy.orm import Session
-from sqlalchemy import text
-from datetime import datetime, timedelta
-from typing import Annotated, Optional
 
-from ...database import SessionLocal, engine, get_db, is_corruption_error, safe_rollback
-from ...models.theme import (
-    ContentSource,
-    ContentItem,
-    ContentItemPipelineState,
-    ThemeMention,
-    ThemeCluster,
-    ThemeConstituent,
-    ThemeMetrics,
-    ThemeAlert,
-    ThemeMergeSuggestion,
-)
-from ...schemas.ui_view_snapshot import UISnapshotEnvelope
-from ...schemas.theme import (
-    ContentSourceCreate,
-    ContentSourceUpdate,
-    ContentSourceResponse,
-    ThemeClusterResponse,
-    ThemeDetailResponse,
-    ThemeConstituentResponse,
-    ThemeRelationshipResponse,
-    ThemeMetricsResponse,
-    ThemeRankingsResponse,
-    ThemeRankingItem,
-    EmergingThemesResponse,
-    EmergingThemeResponse,
-    AlertsResponse,
-    ThemeAlertResponse,
-    ThemeLifecycleTransitionHistoryResponse,
-    ThemeLifecycleTransitionResponse,
-    CandidateThemeQueueResponse,
-    CandidateThemeQueueItemResponse,
-    CandidateThemeQueueSummaryBandResponse,
-    CandidateThemeReviewRequest,
-    CandidateThemeReviewResponse,
-    CandidateThemeReviewItemResult,
-    ThemeMentionsResponse,
-    ThemeMentionDetailResponse,
-    CorrelationDiscoveryResponse,
-    CorrelationClusterResponse,
-    CrossIndustryPairResponse,
-    ThemeValidationResponse,
-    NewEntrantResponse,
-    IngestionResponse,
-    ExtractionResponse,
-    ThemeMergeSuggestionResponse,
-    ThemeMergeSuggestionsResponse,
-    ThemeMergeHistoryListResponse,
-    MergePlanDryRunResponse,
-    MergePlanPairResponse,
-    MergePlanGroupResponse,
-    MergePlanWaveResponse,
-    MergePlanConfidenceBucketResponse,
-    StrictAutoMergeWaveResponse,
-    ManualReviewWaveRequest,
-    ManualReviewWaveResponse,
-    EmbeddingRefreshCampaignResponse,
-    SimilarThemesResponse,
-    SimilarThemeResponse,
-    ThemeRelationshipGraphResponse,
-    ThemeRelationshipGraphNodeResponse,
-    ThemeRelationshipGraphEdgeResponse,
-    ConsolidationResultResponse,
-    MergeActionResponse,
-    ContentItemsListResponse,
-    ContentItemWithThemesResponse,
-    ThemeReference,
-    ThemeMatchBandBucketResponse,
-    ThemeMatchDecisionReasonDistributionResponse,
-    ThemeMatchMethodDistributionResponse,
-    ThemeMatchTelemetryResponse,
-    ThemeMatchTelemetrySliceResponse,
-    ThemePipelineObservabilityResponse,
-    L1ThemeRankingsResponse,
-    L1ThemeRankingItem,
-    L1ChildrenResponse,
-    L1ChildItem,
-    L1ThemeDetail,
-    L1CategoriesResponse,
-    L1CategoryItem,
-    L2ReassignRequest,
-    TaxonomyAssignmentRequest,
-    UnassignedThemesResponse,
-    UnassignedThemeItem,
-    TwitterSessionStatusResponse,
-    TwitterSessionChallengeResponse,
-    TwitterSessionImportRequest,
-)
-from ...services.content_ingestion_service import ContentIngestionService, seed_default_sources
-from ...services.xui_session_bridge_service import (
-    XUISessionBridgeError,
-    XUISessionBridgeService,
-)
-from ...services.theme_extraction_service import ThemeExtractionService
-from ...services.theme_discovery_service import ThemeDiscoveryService
-from ...services.theme_correlation_service import ThemeCorrelationService
-from ...services.theme_merging_service import ThemeMergingService
-from ...services.theme_identity_normalization import UNKNOWN_THEME_KEY, canonical_theme_key, display_theme_name
+from fastapi import APIRouter
+
+from ...database import SessionLocal, is_corruption_error, safe_rollback
+from ...models.theme import ContentItem, ContentItemPipelineState, ContentSource, ThemeMention
 from ...services.theme_content_recovery_service import (
     attempt_reindex_theme_content_storage as _attempt_reindex_theme_content_storage,
     drop_theme_content_tables as _drop_theme_content_tables,
@@ -126,1175 +20,50 @@ from ...services.theme_content_recovery_service import (
     reset_corrupt_theme_content_storage as _reset_corrupt_theme_content_storage,
     rewind_theme_content_source_cursors as _rewind_theme_content_source_cursors,
 )
-from ...services.theme_pipeline_state_service import (
-    compute_pipeline_observability,
-    compute_pipeline_state_health,
-    normalize_pipelines,
-    reconcile_source_pipeline_change,
-    validate_pipeline_selection,
+from ...theme_platform.content_browser_queries import (
+    _build_content_items_browser_base_query as _platform_build_content_items_browser_base_query,
+    _fetch_content_items_with_themes as _platform_fetch_content_items_with_themes,
 )
-from ...wiring.bootstrap import get_ui_snapshot_service
-
-logger = logging.getLogger(__name__)
+from .themes_content_pipeline import router as content_pipeline_router
+from .themes_content_pipeline import get_pipeline_observability
+from .themes_content_sources import router as content_sources_router
+from .themes_content_sources import XUISessionBridgeService
+from .themes_common import (
+    detect_source_type_from_url,
+    resolve_source_ids_for_pipeline as _resolve_source_ids_for_pipeline,
+    safe_theme_cluster_response as _safe_theme_cluster_response,
+)
+from .themes_queries import router as queries_router
+from .themes_queries import (
+    get_lifecycle_transitions,
+    get_matching_telemetry,
+    get_theme_rankings,
+)
+from .themes_review_merge import router as review_merge_router
+from .themes_review_merge import (
+    get_candidate_theme_queue,
+    get_relationship_graph,
+    review_candidate_themes,
+)
+from .themes_taxonomy import router as taxonomy_router
 
 router = APIRouter()
-_VALID_THEME_PIPELINES = {"technical", "fundamental"}
+logger = logging.getLogger(__name__)
 
 
-def detect_source_type_from_url(url: str, provided_type: str | None) -> str:
-    """
-    Auto-detect source type from URL if not provided or if it looks incorrect.
+_build_content_items_browser_base_query = _platform_build_content_items_browser_base_query
+_fetch_content_items_with_themes = _platform_fetch_content_items_with_themes
 
-    Returns the detected source type or the provided type if detection fails.
-    """
-    if not url:
-        return provided_type or "news"
 
-    url_lower = url.lower()
-
-    # Twitter/X detection
-    if 'twitter.com' in url_lower or 'x.com' in url_lower or url_lower.startswith('@'):
-        return 'twitter'
-    # Reddit detection
-    elif 'reddit.com' in url_lower or url_lower.startswith('r/'):
-        return 'reddit'
-    # Substack detection
-    elif 'substack.com' in url_lower:
-        return 'substack'
-    # RSS feed detection
-    elif url_lower.endswith('/feed') or url_lower.endswith('.rss') or url_lower.endswith('.xml'):
-        return 'substack'  # RSS feeds are treated as substack type
-
-    # If provided type exists, use it; otherwise default to news
-    return provided_type or "news"
-
-
-def _normalize_aliases(value: object) -> list[str] | None:
-    if value is None:
-        return None
-    if isinstance(value, list):
-        return [str(alias).strip() for alias in value if str(alias).strip()]
-    alias = str(value).strip()
-    return [alias] if alias else None
-
-
-def _safe_theme_cluster_response(cluster: ThemeCluster) -> ThemeClusterResponse:
-    """Return schema-valid cluster responses, normalizing legacy invalid identity rows."""
-    try:
-        return ThemeClusterResponse.model_validate(cluster)
-    except ValidationError:
-        raw_pipeline = str(getattr(cluster, "pipeline", "") or "").strip().lower()
-        pipeline = raw_pipeline if raw_pipeline in _VALID_THEME_PIPELINES else "technical"
-
-        raw_name = str(getattr(cluster, "name", "") or "").strip()
-        raw_display = str(getattr(cluster, "display_name", "") or "").strip()
-        seed_label = raw_display or raw_name or str(getattr(cluster, "canonical_key", "") or "").strip()
-        display_name = display_theme_name(seed_label)
-        raw_canonical_key = str(getattr(cluster, "canonical_key", "") or "").strip()
-        canonical_key = canonical_theme_key(raw_canonical_key or display_name)
-        if canonical_key == UNKNOWN_THEME_KEY:
-            canonical_key = canonical_theme_key(display_name)
-        name = raw_name or display_name
-
-        logger.warning(
-            "Theme cluster id=%s has invalid identity fields; serving normalized response values",
-            getattr(cluster, "id", None),
-        )
-
-        return ThemeClusterResponse.model_validate(
-            {
-                "id": cluster.id,
-                "name": name,
-                "canonical_key": canonical_key,
-                "display_name": display_name,
-                "aliases": _normalize_aliases(getattr(cluster, "aliases", None)),
-                "description": getattr(cluster, "description", None),
-                "pipeline": pipeline,
-                "category": getattr(cluster, "category", None),
-                "is_emerging": bool(getattr(cluster, "is_emerging", False)),
-                "is_validated": bool(getattr(cluster, "is_validated", False)),
-                "lifecycle_state": str(getattr(cluster, "lifecycle_state", "") or "candidate"),
-                "lifecycle_state_updated_at": getattr(cluster, "lifecycle_state_updated_at", None),
-                "candidate_since_at": getattr(cluster, "candidate_since_at", None),
-                "activated_at": getattr(cluster, "activated_at", None),
-                "dormant_at": getattr(cluster, "dormant_at", None),
-                "reactivated_at": getattr(cluster, "reactivated_at", None),
-                "retired_at": getattr(cluster, "retired_at", None),
-                "discovery_source": getattr(cluster, "discovery_source", None),
-                "first_seen_at": getattr(cluster, "first_seen_at", None),
-                "last_seen_at": getattr(cluster, "last_seen_at", None),
-            }
-        )
-
-
-def _parse_csv_values(value: Optional[str]) -> Optional[list[str]]:
-    if not value:
-        return None
-    values = [item.strip() for item in value.split(",") if item.strip()]
-    return values or None
-
-
-def _resolve_source_ids_for_pipeline(db: Session, pipeline: str) -> list[int]:
-    """Resolve active source IDs assigned to a pipeline."""
-    source_ids: list[int] = []
-    sources = db.query(ContentSource.id, ContentSource.pipelines).filter(
-        ContentSource.is_active == True
-    ).all()
-    for source_id, source_pipelines in sources:
-        if pipeline in normalize_pipelines(source_pipelines):
-            source_ids.append(source_id)
-    return source_ids
-
-
-# ==================== Theme Rankings ====================
-
-@router.get("/bootstrap", response_model=UISnapshotEnvelope)
-async def get_themes_bootstrap(
-    pipeline: str = Query("technical", pattern="^(technical|fundamental)$"),
-    theme_view: str = Query("grouped", pattern="^(grouped|flat)$"),
-    snapshot_service=Depends(get_ui_snapshot_service),
-):
-    """Return the published themes bootstrap snapshot if available."""
-    snapshot = snapshot_service.get_themes_bootstrap(pipeline=pipeline, theme_view=theme_view)
-    if snapshot is None:
-        raise HTTPException(status_code=404, detail="No published themes bootstrap snapshot is available")
-    return UISnapshotEnvelope(**snapshot.to_dict())
-
-@router.get("/pipelines")
-async def get_available_pipelines():
-    """
-    Get list of available theme pipelines.
-
-    Returns technical and fundamental pipeline configurations.
-    """
-    from ...config.pipeline_config import get_all_pipelines
-    return {"pipelines": get_all_pipelines()}
-
-
-@router.get("/rankings", response_model=ThemeRankingsResponse)
-async def get_theme_rankings(
-    limit: int = Query(20, ge=1, le=500, description="Number of themes to return"),
-    offset: int = Query(0, ge=0, description="Number of themes to skip for pagination"),
-    status: Optional[str] = Query(None, description="Filter by status: emerging, trending, fading, dormant"),
-    source_types: Optional[str] = Query(None, description="Comma-separated source types: substack,twitter,news,reddit"),
-    lifecycle_states: Optional[str] = Query(
-        None,
-        description="Comma-separated lifecycle states: candidate,active,dormant,reactivated,retired",
-    ),
-    pipeline: str = Query("technical", description="Pipeline: technical or fundamental"),
-    recalculate: bool = Query(False, description="Force recalculation of metrics"),
-    db: Session = Depends(get_db)
-):
-    """
-    Get current theme rankings sorted by momentum score.
-
-    Returns themes ranked by composite momentum score which combines:
-    - Mention velocity (social signal strength)
-    - Relative strength vs SPY
-    - Breadth (% above 50MA)
-    - Internal correlation (theme cohesiveness)
-    - Screener quality (% passing Minervini)
-
-    Rankings are calculated separately per pipeline.
-    """
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-
-    # Check if metrics need updating (themes without metrics or forced recalc)
-    if recalculate:
-        service.update_all_theme_metrics()
-    else:
-        # Auto-calculate metrics for themes that don't have any
-        from sqlalchemy import func
-        themes_without_metrics = db.query(ThemeCluster).filter(
-            ThemeCluster.is_active == True,
-            ThemeCluster.is_l1 == False,
-            ~ThemeCluster.id.in_(
-                db.query(ThemeMetrics.theme_cluster_id).distinct()
-            )
-        ).count()
-        if themes_without_metrics > 0:
-            logger.info(f"Auto-calculating metrics for {themes_without_metrics} themes without metrics")
-            service.update_all_theme_metrics()
-
-    source_types_list = _parse_csv_values(source_types)
-    lifecycle_states_list = _parse_csv_values(lifecycle_states)
-
-    rankings, total_count = service.get_theme_rankings(
-        limit=limit,
-        status_filter=status,
-        source_types_filter=source_types_list,
-        lifecycle_states_filter=lifecycle_states_list,
-        offset=offset
-    )
-
-    if not rankings:
-        return ThemeRankingsResponse(
-            date=None,
-            total_themes=total_count,
-            pipeline=pipeline,
-            rankings=[]
-        )
-
-    return ThemeRankingsResponse(
-        date=datetime.utcnow().date().isoformat(),
-        total_themes=total_count,
-        pipeline=pipeline,
-        rankings=[ThemeRankingItem(**r) for r in rankings]
-    )
-
-
-@router.get("/emerging", response_model=EmergingThemesResponse)
-async def get_emerging_themes(
-    min_velocity: float = Query(1.5, description="Minimum mention velocity"),
-    min_mentions: int = Query(3, description="Minimum mentions in 7 days"),
-    lifecycle_states: Optional[str] = Query(
-        None,
-        description="Comma-separated lifecycle states: candidate,active,dormant,reactivated,retired",
-    ),
-    pipeline: str = Query("technical", description="Pipeline: technical or fundamental"),
-    db: Session = Depends(get_db)
-):
-    """
-    Discover newly emerging themes.
-
-    Returns themes first seen in last 7 days with accelerating mentions.
-    Filtered by pipeline.
-    """
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-    lifecycle_states_list = _parse_csv_values(lifecycle_states)
-    themes = service.discover_emerging_themes(
-        min_velocity=min_velocity,
-        min_mentions=min_mentions,
-        lifecycle_states_filter=lifecycle_states_list,
-    )
-
-    return EmergingThemesResponse(
-        count=len(themes),
-        themes=[EmergingThemeResponse(**t) for t in themes]
-    )
-
-
-# ==================== Alerts (MUST be before /{theme_id} to avoid route conflict) ====================
-
-@router.get("/alerts", response_model=AlertsResponse)
-async def get_alerts(
-    unread_only: bool = Query(False),
-    limit: int = Query(50, ge=1, le=200),
-    db: Session = Depends(get_db)
-):
-    """
-    Get theme alerts (excludes dismissed alerts).
-    """
-    query = db.query(ThemeAlert).filter(ThemeAlert.is_dismissed == False)
-
-    if unread_only:
-        query = query.filter(ThemeAlert.is_read == False)
-
-    query = query.order_by(ThemeAlert.triggered_at.desc()).limit(limit)
-    alerts = query.all()
-
-    unread_count = db.query(ThemeAlert).filter(
-        ThemeAlert.is_read == False,
-        ThemeAlert.is_dismissed == False
-    ).count()
-
-    return AlertsResponse(
-        total=len(alerts),
-        unread=unread_count,
-        alerts=[ThemeAlertResponse.model_validate(a) for a in alerts]
-    )
-
-
-@router.get("/lifecycle-transitions", response_model=ThemeLifecycleTransitionHistoryResponse)
-async def get_lifecycle_transitions(
-    limit: int = Query(100, ge=1, le=500),
-    offset: int = Query(0, ge=0),
-    pipeline: str = Query("technical", description="Pipeline: technical or fundamental"),
-    theme_cluster_id: Optional[int] = Query(None, description="Optional theme cluster ID filter"),
-    to_state: Optional[str] = Query(
-        None,
-        description="Optional target lifecycle state filter: candidate,active,dormant,reactivated,retired",
-    ),
-    db: Session = Depends(get_db),
-):
-    """
-    Get lifecycle transition audit history with decision context.
-    """
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-    history, total_count = service.get_lifecycle_transition_history(
-        limit=limit,
-        offset=offset,
-        theme_cluster_id=theme_cluster_id,
-        to_state=to_state,
-    )
-
-    return ThemeLifecycleTransitionHistoryResponse(
-        total=total_count,
-        transitions=[ThemeLifecycleTransitionResponse(**row) for row in history],
-    )
-
-
-@router.post("/alerts/{alert_id}/dismiss")
-async def dismiss_alert(
-    alert_id: int,
-    db: Session = Depends(get_db)
-):
-    """Dismiss (soft delete) an alert."""
-    alert = db.query(ThemeAlert).filter(ThemeAlert.id == alert_id).first()
-    if not alert:
-        raise HTTPException(status_code=404, detail="Alert not found")
-    alert.is_dismissed = True
-    db.commit()
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants()
-    return {"status": "dismissed", "alert_id": alert_id}
-
-
-# ==================== Content Sources (MUST be before /{theme_id}) ====================
-
-@router.get("/twitter/session", response_model=TwitterSessionStatusResponse)
-async def get_twitter_session_status() -> TwitterSessionStatusResponse:
-    """Get current XUI auth/session status for twitter ingestion."""
-    service = XUISessionBridgeService()
-    try:
-        status = service.get_auth_status()
-    except XUISessionBridgeError as exc:
-        raise HTTPException(status_code=exc.status_code, detail=exc.detail) from exc
-    return TwitterSessionStatusResponse.model_validate(status, from_attributes=True)
-
-
-@router.post("/twitter/session/challenge", response_model=TwitterSessionChallengeResponse)
-async def create_twitter_session_challenge(request: Request) -> TwitterSessionChallengeResponse:
-    """Issue one-time challenge token for browser-extension session import."""
-    service = XUISessionBridgeService()
-    origin = request.headers.get("origin")
-    client_key = request.client.host if request.client else "unknown"
-    try:
-        challenge = service.create_import_challenge(origin=origin, client_key=client_key)
-    except XUISessionBridgeError as exc:
-        raise HTTPException(status_code=exc.status_code, detail=exc.detail) from exc
-    return TwitterSessionChallengeResponse.model_validate(challenge, from_attributes=True)
-
-
-@router.post("/twitter/session/import", response_model=TwitterSessionStatusResponse)
-async def import_twitter_session_from_browser(
-    payload: TwitterSessionImportRequest,
-    request: Request,
-) -> TwitterSessionStatusResponse:
-    """Import x.com/twitter.com cookies from extension and persist XUI storage state."""
-    service = XUISessionBridgeService()
-    origin = request.headers.get("x-xui-bridge-origin") or request.headers.get("origin")
-    client_key = request.client.host if request.client else "unknown"
-    try:
-        status = service.import_browser_cookies(
-            challenge_id=payload.challenge_id,
-            challenge_token=payload.challenge_token,
-            cookies=[cookie.model_dump() for cookie in payload.cookies],
-            origin=origin,
-            client_key=client_key,
-            browser=payload.browser,
-            extension_version=payload.extension_version,
-        )
-    except XUISessionBridgeError as exc:
-        raise HTTPException(status_code=exc.status_code, detail=exc.detail) from exc
-    return TwitterSessionStatusResponse.model_validate(status, from_attributes=True)
-
-@router.get("/sources", response_model=list[ContentSourceResponse])
-async def list_content_sources(
-    active_only: bool = Query(True),
-    pipeline: Optional[str] = Query(None, description="Filter by pipeline: technical or fundamental"),
-    db: Session = Depends(get_db)
-):
-    """
-    List all configured content sources.
-
-    Optionally filter by pipeline assignment.
-    """
-    query = db.query(ContentSource)
-    if active_only:
-        query = query.filter(ContentSource.is_active == True)
-
-    sources = query.order_by(ContentSource.priority.desc()).all()
-
-    # Filter by pipeline if specified
-    if pipeline:
-        import json
-        filtered_sources = []
-        for source in sources:
-            source_pipelines = source.pipelines or ["technical", "fundamental"]
-            if isinstance(source_pipelines, str):
-                try:
-                    source_pipelines = json.loads(source_pipelines)
-                except:
-                    source_pipelines = ["technical", "fundamental"]
-            if pipeline in source_pipelines:
-                filtered_sources.append(source)
-        sources = filtered_sources
-
-    return [ContentSourceResponse.model_validate(s) for s in sources]
-
-
-@router.post("/sources", response_model=ContentSourceResponse)
-async def add_content_source(
-    source: ContentSourceCreate,
-    db: Session = Depends(get_db)
-):
-    """
-    Add a new content source for theme extraction.
-
-    The pipelines field determines which pipelines this source feeds into.
-    Source type is auto-detected from URL if it appears incorrect (e.g., Twitter URL with substack type).
-    """
-    # Auto-detect source type from URL/name if it looks like it might be wrong
-    detected_type = detect_source_type_from_url(source.url or source.name, source.source_type)
-
-    # Log if we're correcting the source type
-    if detected_type != source.source_type:
-        logger.info(f"Auto-correcting source type for '{source.name}': {source.source_type} -> {detected_type}")
-
-    try:
-        pipelines = validate_pipeline_selection(source.pipelines)
-    except ValueError as exc:
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    # Create source directly since ContentIngestionService.add_source may not support pipelines yet
-    new_source = ContentSource(
-        name=source.name,
-        source_type=detected_type,
-        url=source.url,
-        priority=source.priority,
-        fetch_interval_minutes=source.fetch_interval_minutes,
-        pipelines=pipelines,
-        is_active=True,
-        total_items_fetched=0,
-    )
-    db.add(new_source)
-    db.commit()
-    db.refresh(new_source)
-    return ContentSourceResponse.model_validate(new_source)
-
-
-@router.put("/sources/{source_id}", response_model=ContentSourceResponse)
-async def update_content_source(
-    source_id: int,
-    source: ContentSourceUpdate,
-    db: Session = Depends(get_db)
-):
-    """
-    Update an existing content source.
-
-    When the pipelines field changes, existing ThemeMention records are preserved.
-    Pipeline-state rows are reconciled non-destructively for added/removed
-    assignments.
-    """
-    existing = db.query(ContentSource).filter(ContentSource.id == source_id).first()
-    if not existing:
-        raise HTTPException(status_code=404, detail="Source not found")
-
-    # Capture old pipelines BEFORE update for comparison
-    old_pipelines = normalize_pipelines(existing.pipelines)
-    new_pipelines = old_pipelines
-
-    # Update fields if provided
-    if source.name is not None:
-        existing.name = source.name
-    if source.source_type is not None:
-        existing.source_type = source.source_type
-    if source.url is not None:
-        existing.url = source.url
-    if source.priority is not None:
-        existing.priority = source.priority
-    if source.fetch_interval_minutes is not None:
-        existing.fetch_interval_minutes = source.fetch_interval_minutes
-    if source.is_active is not None:
-        existing.is_active = source.is_active
-    try:
-        if source.pipelines is not None:
-            new_pipelines = validate_pipeline_selection(source.pipelines)
-            existing.pipelines = new_pipelines
-
-        # Ensure source update + reconciliation complete atomically for this request.
-        db.flush()
-
-        if source.pipelines is not None and set(old_pipelines) != set(new_pipelines):
-            reconcile_summary = reconcile_source_pipeline_change(
-                db=db,
-                source_id=source_id,
-                old_pipelines=old_pipelines,
-                new_pipelines=new_pipelines,
-                commit_each_chunk=False,
-            )
-            logger.info("Source %s pipeline change reconciled: %s", source_id, reconcile_summary)
-
-        db.commit()
-        db.refresh(existing)
-    except ValueError as exc:
-        db.rollback()
-        raise HTTPException(status_code=400, detail=str(exc)) from exc
-    except Exception:
-        db.rollback()
-        raise
-
-    return ContentSourceResponse.model_validate(existing)
-
-
-@router.delete("/sources/{source_id}")
-async def delete_content_source(
-    source_id: int,
-    db: Session = Depends(get_db)
-):
-    """
-    Deactivate a content source.
-    """
-    source = db.query(ContentSource).filter(ContentSource.id == source_id).first()
-    if not source:
-        raise HTTPException(status_code=404, detail="Source not found")
-
-    source.is_active = False
-    db.commit()
-
-    return {"status": "deactivated", "source": source.name}
-
-
-@router.post("/sources/seed-defaults")
-async def seed_default_sources_endpoint(
-    db: Session = Depends(get_db)
-):
-    """
-    Seed the database with default content sources.
-    """
-    seed_default_sources(db)
-    return {"status": "success", "message": "Default sources seeded"}
-
-
-# ==================== Ingestion & Extraction (MUST be before /{theme_id}) ====================
-
-@router.post("/ingest", response_model=IngestionResponse)
-async def run_ingestion(
-    background_tasks: BackgroundTasks,
-    db: Session = Depends(get_db)
-):
-    """
-    Fetch new content from all active sources.
-
-    This pulls content from RSS feeds, Twitter, news APIs, etc.
-    """
-    service = ContentIngestionService(db)
-    result = service.fetch_all_active_sources()
-
-    return IngestionResponse(**result)
-
-
-@router.post("/extract", response_model=ExtractionResponse)
-async def run_extraction(
-    limit: int = Query(50, ge=1, le=200, description="Max items to process"),
-    pipeline: str = Query("technical", description="Pipeline: technical or fundamental"),
-    db: Session = Depends(get_db)
-):
-    """
-    Extract themes from unprocessed content using LLM.
-
-    Processes content items that haven't been analyzed yet,
-    extracting themes, tickers, and sentiment.
-
-    Pipeline parameter determines which content sources to process and
-    which pipeline-specific extraction prompt to use.
-    """
-    service = ThemeExtractionService(db, pipeline=pipeline)
-    result = service.process_batch(limit=limit)
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants(pipeline)
-
-    return ExtractionResponse(**result)
-
-
-@router.post("/calculate-metrics")
-async def calculate_theme_metrics(
-    pipeline: str = Query("technical", description="Pipeline: technical or fundamental"),
-    db: Session = Depends(get_db)
-):
-    """
-    Calculate/update metrics for all active themes in a pipeline.
-
-    This updates mention velocity, price performance, correlations,
-    and rankings for all themes in the specified pipeline.
-    Uses pipeline-specific scoring weights.
-    """
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-    result = service.update_all_theme_metrics()
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants(pipeline)
-
-    return result
-
-
-@router.post("/validate-all")
-async def validate_all_themes(
-    min_correlation: float = Query(0.5, ge=0.2, le=0.9),
-    db: Session = Depends(get_db)
-):
-    """
-    Run validation on all active themes.
-
-    Checks internal correlations and identifies invalid themes.
-    """
-    service = ThemeCorrelationService(db)
-    result = service.run_full_validation(min_correlation)
-
-    return result
-
-
-# ==================== Async Pipeline (Celery-based) ====================
-
-@router.post("/pipeline/run")
-async def run_pipeline_async(
-    pipeline: Optional[str] = Query(None, description="Pipeline: technical, fundamental, or None for both"),
-    lookback_days: Optional[int] = Query(None, ge=1, le=30, description="Re-fetch articles from the last N days (backfill mode)"),
-    db: Session = Depends(get_db)
-):
-    """
-    Start the full theme discovery pipeline asynchronously.
-
-    Queues a Celery task to run:
-    1. Content ingestion from all active sources
-    2. Reprocess previously failed extractions (retry)
-    3. Theme extraction via LLM (with pipeline-specific prompts)
-    4. Metrics calculation for all themes (with pipeline-specific weights)
-    5. Alert generation
-
-    If pipeline is None, runs for both technical and fundamental sequentially.
-
-    Returns a run_id for polling status.
-    """
-    import uuid
-    from ...tasks.theme_discovery_tasks import run_full_pipeline
-    from ...models.theme import ThemePipelineRun
-
-    # Generate unique run ID
-    run_id = str(uuid.uuid4())
-
-    # Queue Celery task with pipeline parameter
-    task = run_full_pipeline.delay(run_id=run_id, pipeline=pipeline, lookback_days=lookback_days)
-
-    # Create pipeline run record for tracking
-    pipeline_run = ThemePipelineRun(
-        run_id=run_id,
-        task_id=task.id,
-        pipeline=pipeline,
-        status="queued"
-    )
-    db.add(pipeline_run)
-    db.commit()
-
-    pipeline_desc = pipeline if pipeline else "both (technical + fundamental)"
-    logger.info(f"Theme pipeline {run_id} queued for {pipeline_desc} with task ID: {task.id}")
-
-    return {
-        "run_id": run_id,
-        "task_id": task.id,
-        "status": "queued",
-        "pipeline": pipeline,
-        "message": f"Theme discovery pipeline queued for {pipeline_desc}"
-    }
-
-
-@router.get("/pipeline/{run_id}/status")
-async def get_pipeline_status(
-    run_id: str,
-    db: Session = Depends(get_db)
-):
-    """
-    Get current status of a pipeline run.
-
-    Polls Celery task state for real-time progress updates.
-    Falls back to database record if task state unavailable.
-    """
-    from celery.result import AsyncResult
-    from ...models.theme import ThemePipelineRun
-
-    # Get pipeline run from database
-    pipeline_run = db.query(ThemePipelineRun).filter(
-        ThemePipelineRun.run_id == run_id
-    ).first()
-
-    if not pipeline_run:
-        raise HTTPException(status_code=404, detail=f"Pipeline run {run_id} not found")
-
-    # Default response from DB
-    response = {
-        'run_id': run_id,
-        'task_id': pipeline_run.task_id,
-        'status': pipeline_run.status,
-        'current_step': pipeline_run.current_step,
-        'step_number': 0,
-        'total_steps': 5,
-        'percent': 0.0,
-        'message': None,
-        'ingestion_result': None,
-        'reprocessing_result': None,
-        'extraction_result': None,
-        'metrics_result': None,
-        'alerts_result': None,
-        'started_at': pipeline_run.started_at.isoformat() if pipeline_run.started_at else None,
-        'completed_at': pipeline_run.completed_at.isoformat() if pipeline_run.completed_at else None,
-        'error_message': pipeline_run.error_message,
-    }
-
-    # If completed or failed, return DB state
-    if pipeline_run.status in ['completed', 'failed']:
-        if pipeline_run.status == 'completed':
-            response['percent'] = 100.0
-            response['step_number'] = 5
-            response['current_step'] = 'completed'
-        return response
-
-    # Get real-time progress from Celery task
-    if pipeline_run.task_id:
-        task_result = AsyncResult(pipeline_run.task_id)
-
-        if task_result.state == 'PROGRESS' and task_result.info:
-            info = task_result.info
-            response['current_step'] = info.get('current_step')
-            response['step_number'] = info.get('step_number', 0)
-            response['percent'] = info.get('percent', 0.0)
-            response['message'] = info.get('message')
-            response['ingestion_result'] = info.get('ingestion_result')
-            response['reprocessing_result'] = info.get('reprocessing_result')
-            response['extraction_result'] = info.get('extraction_result')
-            response['metrics_result'] = info.get('metrics_result')
-            response['status'] = 'running'
-        elif task_result.state == 'SUCCESS':
-            response['status'] = 'completed'
-            response['percent'] = 100.0
-            response['step_number'] = 5
-            response['current_step'] = 'completed'
-            if task_result.result:
-                response['alerts_result'] = task_result.result.get('steps', {}).get('alerts')
-        elif task_result.state == 'FAILURE':
-            response['status'] = 'failed'
-            response['error_message'] = str(task_result.result) if task_result.result else 'Unknown error'
-
-    return response
-
-
-@router.get("/pipeline/runs")
-async def list_pipeline_runs(
-    limit: int = Query(10, ge=1, le=50),
-    db: Session = Depends(get_db)
-):
-    """
-    List recent pipeline runs.
-    """
-    from ...models.theme import ThemePipelineRun
-
-    runs = db.query(ThemePipelineRun).order_by(
-        ThemePipelineRun.started_at.desc()
-    ).limit(limit).all()
-
-    return {
-        'total': len(runs),
-        'runs': [
-            {
-                'run_id': r.run_id,
-                'status': r.status,
-                'current_step': r.current_step,
-                'items_ingested': r.items_ingested,
-                'items_reprocessed': r.items_reprocessed,
-                'themes_updated': r.themes_updated,
-                'started_at': r.started_at.isoformat() if r.started_at else None,
-                'completed_at': r.completed_at.isoformat() if r.completed_at else None,
-            }
-            for r in runs
-        ]
-    }
-
-
-@router.get("/pipeline/failed-count")
-async def get_failed_items_count(
-    pipeline: Optional[str] = Query(None, description="Filter by pipeline"),
-    db: Session = Depends(get_db),
-):
-    """
-    Count content items with extraction errors eligible for reprocessing.
-
-    Returns the number of items that will be retried on the next pipeline run.
-    """
-    from datetime import timedelta
-    from sqlalchemy import func
-
-    cutoff = datetime.utcnow() - timedelta(days=30)
-
-    query = db.query(func.count(ContentItemPipelineState.id)).join(
-        ContentItem,
-        ContentItem.id == ContentItemPipelineState.content_item_id,
-    ).filter(
-        ContentItemPipelineState.status == "failed_retryable",
-        ContentItem.published_at >= cutoff,
-    )
-
-    # Filter by pipeline and active source assignments if specified
-    if pipeline:
-        if pipeline not in {"technical", "fundamental"}:
-            raise HTTPException(status_code=400, detail="pipeline must be technical or fundamental")
-
-        query = query.filter(ContentItemPipelineState.pipeline == pipeline)
-        import json as json_lib
-        source_ids = []
-        sources = db.query(ContentSource).filter(ContentSource.is_active == True).all()
-        for source in sources:
-            source_pipelines = source.pipelines or ["technical", "fundamental"]
-            if isinstance(source_pipelines, str):
-                try:
-                    source_pipelines = json_lib.loads(source_pipelines)
-                except Exception:
-                    source_pipelines = ["technical", "fundamental"]
-            if pipeline in source_pipelines:
-                source_ids.append(source.id)
-        if source_ids:
-            query = query.filter(ContentItem.source_id.in_(source_ids))
-        else:
-            return {"failed_count": 0, "max_age_days": 30}
-
-    count = query.scalar()
-    return {"failed_count": count, "max_age_days": 30}
-
-
-@router.get("/pipeline/state-health")
-async def get_pipeline_state_health(
-    pipeline: Optional[str] = Query(None, description="Pipeline: technical or fundamental"),
-    window_days: int = Query(30, ge=1, le=365, description="Lookback window in days"),
-    db: Session = Depends(get_db),
-):
-    """
-    Pipeline-state health and drift metrics for operational triage.
-
-    Includes:
-    - pending/in-progress/processed/retryable/terminal counts
-    - pending age percentiles
-    - retry queue growth (last 24h vs previous 24h)
-    - parse-failure rate
-    - processed-without-mentions ratio
-    """
-    if pipeline and pipeline not in {"technical", "fundamental"}:
-        raise HTTPException(status_code=400, detail="pipeline must be technical or fundamental")
-
-    return compute_pipeline_state_health(
-        db=db,
-        pipeline=pipeline,
-        max_age_days=window_days,
-    )
-
-
-@router.get("/pipeline/observability", response_model=ThemePipelineObservabilityResponse)
-async def get_pipeline_observability(
-    pipeline: str = Query(..., description="Pipeline: technical or fundamental"),
-    window_days: int = Query(30, ge=1, le=365, description="Lookback window in days"),
-    db: Session = Depends(get_db),
-):
-    """
-    Dashboard summary + actionable alert policies for pipeline health.
-    """
-    if pipeline not in {"technical", "fundamental"}:
-        raise HTTPException(status_code=400, detail="pipeline must be technical or fundamental")
-    return compute_pipeline_observability(
-        db=db,
-        pipeline=pipeline,
-        max_age_days=window_days,
-    )
-
-
-# ==================== Content Item Browser (MUST be before /{theme_id}) ====================
-
-
-def _build_content_items_browser_base_query(
-    db: Session,
-    *,
-    source_type: Optional[str] = None,
-    date_from: Optional[str] = None,
-    date_to: Optional[str] = None,
-    pipeline: Optional[str] = None,
-    pipeline_source_ids: Optional[list[int]] = None,
-):
-    """Build the shared base query for content browser list/export reads."""
-    from datetime import datetime as dt, timedelta
-
-    base_query = db.query(ContentItem).join(
-        ContentSource, ContentItem.source_id == ContentSource.id
-    ).filter(
-        ContentSource.is_active == True
-    )
-
-    if pipeline:
-        if pipeline_source_ids is None:
-            pipeline_source_ids = _resolve_source_ids_for_pipeline(db, pipeline)
-        if not pipeline_source_ids:
-            return None
-        base_query = base_query.filter(ContentItem.source_id.in_(pipeline_source_ids))
-    else:
-        base_query = base_query.filter(ContentItem.is_processed == True)
-
-    if source_type:
-        base_query = base_query.filter(ContentItem.source_type == source_type)
-
-    if date_from:
-        try:
-            from_date = dt.strptime(date_from, "%Y-%m-%d")
-            base_query = base_query.filter(ContentItem.published_at >= from_date)
-        except ValueError:
-            pass
-
-    if date_to:
-        try:
-            to_date = dt.strptime(date_to, "%Y-%m-%d") + timedelta(days=1)
-            base_query = base_query.filter(ContentItem.published_at < to_date)
-        except ValueError:
-            pass
-
-    return base_query
-
-
-def _fetch_content_items_with_themes(
-    db: Session,
-    search: Optional[str] = None,
-    source_type: Optional[str] = None,
-    sentiment: Optional[str] = None,
-    date_from: Optional[str] = None,
-    date_to: Optional[str] = None,
-    sort_by: str = "published_at",
-    sort_order: str = "desc",
-    limit: Optional[int] = None,
-    offset: Optional[int] = None,
-    pipeline: Optional[str] = None,
-) -> tuple[list[ContentItemWithThemesResponse], int]:
-    """
-    Shared query logic for content items with theme/sentiment/ticker aggregation.
-
-    Returns (items, total_count) where items are ContentItemWithThemesResponse objects.
-    When limit/offset are None, returns all matching items (used by export).
-    """
-    from sqlalchemy import or_, desc, asc, func
-
-    # Base query defaults to the historical processed-only behavior.
-    # With a pipeline provided, scope to assigned sources and include all fetched items.
-    base_query = _build_content_items_browser_base_query(
-        db,
-        source_type=source_type,
-        date_from=date_from,
-        date_to=date_to,
-        pipeline=pipeline,
-    )
-    if base_query is None:
-        return [], 0
-
-    # Apply sentiment filter before pagination/count so pages are accurate.
-    if sentiment:
-        sentiment_query = db.query(ThemeMention.content_item_id).filter(
-            ThemeMention.content_item_id.in_(base_query.with_entities(ContentItem.id))
-        )
-        if pipeline:
-            sentiment_query = sentiment_query.filter(ThemeMention.pipeline == pipeline)
-        sentiment_content_ids = {
-            row.content_item_id for row in sentiment_query.filter(
-                ThemeMention.sentiment.isnot(None),
-                func.lower(ThemeMention.sentiment) == sentiment.lower(),
-            ).distinct().all()
-        }
-        if not sentiment_content_ids:
-            return [], 0
-        base_query = base_query.filter(ContentItem.id.in_(sentiment_content_ids))
-
-    # Search applies to title/source OR mentioned ticker, both before pagination.
-    if search:
-        search_term = f"%{search}%"
-        text_match_query = base_query.with_entities(ContentItem.id).filter(
-            or_(
-                ContentItem.title.ilike(search_term),
-                ContentItem.source_name.ilike(search_term),
-            )
-        )
-        text_matched_ids = {row.id for row in text_match_query.all()}
-
-        ticker_match_query = db.query(
-            ThemeMention.content_item_id,
-            ThemeMention.tickers,
-        ).filter(
-            ThemeMention.content_item_id.in_(base_query.with_entities(ContentItem.id))
-        )
-        if pipeline:
-            ticker_match_query = ticker_match_query.filter(ThemeMention.pipeline == pipeline)
-
-        search_upper = search.upper()
-        ticker_matched_ids: set[int] = set()
-        for mention in ticker_match_query.all():
-            if mention.tickers:
-                for ticker in mention.tickers:
-                    if search_upper in str(ticker).upper():
-                        ticker_matched_ids.add(mention.content_item_id)
-                        break
-
-        matched_ids = text_matched_ids | ticker_matched_ids
-        if not matched_ids:
-            return [], 0
-        base_query = base_query.filter(ContentItem.id.in_(matched_ids))
-
-    # Get total count before pagination
-    total = base_query.count()
-
-    # Apply sorting
-    sort_column = getattr(ContentItem, sort_by, ContentItem.published_at)
-    if sort_order.lower() == "asc":
-        base_query = base_query.order_by(asc(sort_column))
-    else:
-        base_query = base_query.order_by(desc(sort_column))
-
-    # Apply pagination only if provided
-    if limit is not None and offset is not None:
-        content_items = base_query.offset(offset).limit(limit).all()
-    else:
-        content_items = base_query.all()
-
-    if not content_items:
-        return [], total
-
-    # Get all content item IDs for batch fetching mentions
-    content_ids = [item.id for item in content_items]
-
-    # Fetch all theme mentions for these content items
-    mentions = db.query(
-        ThemeMention.content_item_id,
-        ThemeMention.theme_cluster_id,
-        ThemeMention.sentiment,
-        ThemeMention.tickers,
-        ThemeCluster.id.label("cluster_id"),
-        ThemeCluster.display_name.label("cluster_name")
-    ).outerjoin(
-        ThemeCluster, ThemeMention.theme_cluster_id == ThemeCluster.id
-    ).filter(
-        ThemeMention.content_item_id.in_(content_ids)
-    )
-    if pipeline:
-        mentions = mentions.filter(ThemeMention.pipeline == pipeline)
-    mentions = mentions.all()
-
-    pipeline_status_by_content_id: dict[int, str] = {}
-    if pipeline:
-        status_rows = db.query(
-            ContentItemPipelineState.content_item_id,
-            ContentItemPipelineState.status,
-        ).filter(
-            ContentItemPipelineState.content_item_id.in_(content_ids),
-            ContentItemPipelineState.pipeline == pipeline,
-        ).all()
-        pipeline_status_by_content_id = {
-            row.content_item_id: row.status for row in status_rows
-        }
-
-    # Aggregate mentions by content_item_id
-    mentions_by_content = {}
-    for mention in mentions:
-        content_id = mention.content_item_id
-        if content_id not in mentions_by_content:
-            mentions_by_content[content_id] = {
-                "themes": [],
-                "sentiments": [],
-                "tickers": set()
-            }
-
-        if mention.cluster_id and mention.cluster_name:
-            theme_ref = {"id": mention.cluster_id, "name": mention.cluster_name}
-            if theme_ref not in mentions_by_content[content_id]["themes"]:
-                mentions_by_content[content_id]["themes"].append(theme_ref)
-
-        if mention.sentiment:
-            if mention.sentiment not in mentions_by_content[content_id]["sentiments"]:
-                mentions_by_content[content_id]["sentiments"].append(mention.sentiment)
-
-        if mention.tickers:
-            mentions_by_content[content_id]["tickers"].update(mention.tickers)
-
-    # Build response items
-    items = []
-    for content in content_items:
-        mention_data = mentions_by_content.get(content.id, {"themes": [], "sentiments": [], "tickers": set()})
-
-        sentiments_list = mention_data["sentiments"]
-        primary_sentiment = None
-        if sentiments_list:
-            sentiment_counts = {}
-            for s in sentiments_list:
-                sentiment_counts[s] = sentiment_counts.get(s, 0) + 1
-            primary_sentiment = max(sentiment_counts, key=sentiment_counts.get)
-
-        items.append(ContentItemWithThemesResponse(
-            id=content.id,
-            title=content.title,
-            content=content.content,
-            url=content.url,
-            source_type=content.source_type,
-            source_name=content.source_name,
-            author=content.author,
-            published_at=content.published_at,
-            themes=[ThemeReference(**t) for t in mention_data["themes"]],
-            sentiments=sentiments_list,
-            primary_sentiment=primary_sentiment,
-            tickers=sorted(list(mention_data["tickers"])),
-            processing_status=(
-                pipeline_status_by_content_id.get(content.id, "pending")
-                if pipeline
-                else None
-            ),
-        ))
-
-    return items, total
-
-
-def _fetch_content_items_with_themes_with_recovery(
-    db: Session,
-    **kwargs,
-) -> tuple[list[ContentItemWithThemesResponse], int]:
-    """Retry theme content queries after reindexing or rebuilding recoverable storage."""
-    try:
-        return _fetch_content_items_with_themes(db, **kwargs)
-    except Exception as exc:
-        if not is_corruption_error(exc):
-            raise
-        safe_rollback(db)
-        _attempt_reindex_theme_content_storage(exc)
-        try:
-            with SessionLocal() as retry_db:
-                return _fetch_content_items_with_themes(retry_db, **kwargs)
-        except Exception as retry_exc:
-            if not is_corruption_error(retry_exc):
-                raise
-            logger.warning(
-                "Theme content browser query still hits corruption after REINDEX; "
-                "checking whether rebuildable theme-content storage can be reset: %s",
-                retry_exc,
-            )
-            if not _corruption_targets_theme_content_storage(**kwargs):
-                logger.error(
-                    "Theme content browser detected database corruption outside rebuildable "
-                    "theme-content storage. Run backend/scripts/check_db_integrity.py --repair "
-                    "or restore the latest valid backup. Initial error: %s. Retry error: %s",
-                    exc,
-                    retry_exc,
-                )
-                raise
-            _reset_corrupt_theme_content_storage(retry_exc)
-            with SessionLocal() as reset_retry_db:
-                return _fetch_content_items_with_themes(reset_retry_db, **kwargs)
 def _corruption_targets_theme_content_storage(
     *,
-    source_type: Optional[str] = None,
-    date_from: Optional[str] = None,
-    date_to: Optional[str] = None,
-    pipeline: Optional[str] = None,
+    source_type=None,
+    date_from=None,
+    date_to=None,
+    pipeline=None,
     **_ignored,
-) -> bool:
-    """Return True when filtered browser probes hit rebuildable theme-content storage corruption."""
+):
+    """Compatibility wrapper that preserves monkeypatchable classifier probes."""
     with SessionLocal() as probe_db:
         try:
             probe_db.query(ContentSource.id).filter(
@@ -1368,1143 +137,35 @@ def _corruption_targets_theme_content_storage(
     return False
 
 
-@router.get("/content", response_model=ContentItemsListResponse)
-async def list_content_items(
-    search: Optional[str] = Query(None, description="Search in title, source_name, tickers"),
-    source_type: Optional[str] = Query(None, description="Filter: substack, twitter, news, reddit"),
-    sentiment: Optional[str] = Query(None, description="Filter: bullish, bearish, neutral"),
-    date_from: Optional[str] = Query(None, description="From date (YYYY-MM-DD)"),
-    date_to: Optional[str] = Query(None, description="To date (YYYY-MM-DD)"),
-    pipeline: Optional[str] = Query(None, description="Pipeline: technical or fundamental"),
-    limit: int = Query(50, ge=1, le=200, description="Items per page"),
-    offset: int = Query(0, ge=0, description="Pagination offset"),
-    sort_by: str = Query("published_at", description="Sort column"),
-    sort_order: str = Query("desc", description="asc or desc"),
-    db: Session = Depends(get_db)
-):
-    """
-    List all content items with their associated themes, sentiments, and tickers.
-
-    Aggregates data from ContentItem and ThemeMention tables to provide a unified
-    view of all ingested content with their extracted metadata.
-    """
-    if pipeline and pipeline not in _VALID_THEME_PIPELINES:
-        raise HTTPException(status_code=400, detail="pipeline must be technical or fundamental")
-
-    items, total = _fetch_content_items_with_themes_with_recovery(
-        db, search=search, source_type=source_type, sentiment=sentiment,
-        date_from=date_from, date_to=date_to, sort_by=sort_by,
-        sort_order=sort_order, limit=limit, offset=offset, pipeline=pipeline,
-    )
-
-    return ContentItemsListResponse(
-        total=total,
-        limit=limit,
-        offset=offset,
-        items=items
-    )
-
-
-@router.get("/content/export")
-async def export_content_items(
-    search: Optional[str] = Query(None, description="Search in title, source_name, tickers"),
-    source_type: Optional[str] = Query(None, description="Filter: substack, twitter, news, reddit"),
-    sentiment: Optional[str] = Query(None, description="Filter: bullish, bearish, neutral"),
-    date_from: Optional[str] = Query(None, description="From date (YYYY-MM-DD)"),
-    date_to: Optional[str] = Query(None, description="To date (YYYY-MM-DD)"),
-    pipeline: Optional[str] = Query(None, description="Pipeline: technical or fundamental"),
-    sort_by: str = Query("published_at", description="Sort column"),
-    sort_order: str = Query("desc", description="asc or desc"),
-    db: Session = Depends(get_db)
-):
-    """
-    Export all content items matching filters as a CSV file.
-
-    Returns all matching items (no pagination) with full article body text.
-    CSV includes: ID, Title, Content, URL, Themes, Sentiment, Tickers,
-    Published Date, Source Type, Source Name, Author.
-    """
-    if pipeline and pipeline not in _VALID_THEME_PIPELINES:
-        raise HTTPException(status_code=400, detail="pipeline must be technical or fundamental")
-
-    items, total = _fetch_content_items_with_themes_with_recovery(
-        db, search=search, source_type=source_type, sentiment=sentiment,
-        date_from=date_from, date_to=date_to, sort_by=sort_by,
-        sort_order=sort_order, pipeline=pipeline,
-    )
-
-    # Build CSV in memory
-    buf = io.StringIO()
-    writer = csv.writer(buf)
-
-    # Header row
-    writer.writerow([
-        "ID", "Title", "Content", "URL", "Themes", "Sentiment",
-        "Tickers", "Published Date", "Source Type", "Source Name", "Author", "Processing Status",
-    ])
-
-    # Data rows
-    for item in items:
-        writer.writerow([
-            item.id,
-            item.title or "",
-            item.content or "",
-            item.url or "",
-            "; ".join(t.name for t in item.themes),
-            item.primary_sentiment or "",
-            "; ".join(item.tickers),
-            item.published_at.strftime("%Y-%m-%d %H:%M") if item.published_at else "",
-            item.source_type or "",
-            item.source_name or "",
-            item.author or "",
-            item.processing_status or "",
-        ])
-
-    # UTF-8 BOM for Excel compatibility
-    csv_bytes = b"\xef\xbb\xbf" + buf.getvalue().encode("utf-8")
-    filename = f"theme_articles_{datetime.utcnow().strftime('%Y%m%d')}.csv"
-
-    return StreamingResponse(
-        io.BytesIO(csv_bytes),
-        media_type="text/csv",
-        headers={"Content-Disposition": f'attachment; filename="{filename}"'},
-    )
-
-
-# ==================== Theme Merging (MUST be before /{theme_id}) ====================
-
-@router.get("/merge-suggestions", response_model=ThemeMergeSuggestionsResponse)
-async def get_merge_suggestions(
-    status: Optional[str] = Query(None, description="Filter by status: pending, approved, rejected, auto_merged"),
-    limit: int = Query(50, ge=1, le=200),
-    db: Session = Depends(get_db)
-):
-    """
-    Get theme merge suggestions from the queue.
-
-    Returns pairs of themes that may be duplicates, with embedding similarity
-    scores and LLM verification results.
-
-    Canonical response fields:
-    - source_theme_id / source_theme_name / target_theme_id / target_theme_name
-    - similarity_score / relationship_type / reasoning / suggested_name
-    Legacy merge fields remain temporarily for migration compatibility.
-    """
-    service = ThemeMergingService(db)
-    suggestions = service.get_merge_suggestions(status=status, limit=limit)
-
-    return ThemeMergeSuggestionsResponse(
-        total=len(suggestions),
-        suggestions=[ThemeMergeSuggestionResponse(**s) for s in suggestions]
-    )
-
-
-@router.post("/merge-suggestions/{suggestion_id}/approve", response_model=MergeActionResponse)
-async def approve_merge_suggestion(
-    suggestion_id: int,
-    idempotency_key: Optional[str] = Header(None, alias="Idempotency-Key"),
-    db: Session = Depends(get_db)
-):
-    """
-    Approve a merge suggestion and execute the merge.
-
-    The source theme will be merged INTO the target theme:
-    - Aliases are combined
-    - Constituents are reassigned
-    - Mentions are reassigned
-    - Source theme is deactivated
-    """
-    suggestion = db.query(ThemeMergeSuggestion).filter(ThemeMergeSuggestion.id == suggestion_id).first()
-    source_pipeline = None
-    if suggestion is not None:
-        source_pipeline = db.query(ThemeCluster.pipeline).filter(
-            ThemeCluster.id == suggestion.source_cluster_id
-        ).scalar()
-    service = ThemeMergingService(db)
-    result = service.approve_suggestion(suggestion_id, idempotency_key=idempotency_key)
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants(source_pipeline)
-    return MergeActionResponse(**result)
-
-
-@router.post("/merge-suggestions/{suggestion_id}/reject")
-async def reject_merge_suggestion(
-    suggestion_id: int,
-    db: Session = Depends(get_db)
-):
-    """
-    Reject a merge suggestion.
-
-    The suggestion is marked as rejected and will not be suggested again.
-    """
-    suggestion = db.query(ThemeMergeSuggestion).filter(ThemeMergeSuggestion.id == suggestion_id).first()
-    source_pipeline = None
-    if suggestion is not None:
-        source_pipeline = db.query(ThemeCluster.pipeline).filter(
-            ThemeCluster.id == suggestion.source_cluster_id
-        ).scalar()
-    service = ThemeMergingService(db)
-    result = service.reject_suggestion(suggestion_id)
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants(source_pipeline)
-    return result
-
-
-@router.get("/merge-history", response_model=ThemeMergeHistoryListResponse)
-async def get_merge_history(
-    limit: int = Query(50, ge=1, le=200),
-    db: Session = Depends(get_db)
-):
-    """
-    Get history of theme merges.
-
-    Returns audit trail of all merges (auto and manual).
-    """
-    service = ThemeMergingService(db)
-    history = service.get_merge_history(limit=limit)
-
-    return ThemeMergeHistoryListResponse(
-        total=len(history),
-        history=history
-    )
-
-
-@router.post("/consolidate", response_model=ConsolidationResultResponse)
-async def run_theme_consolidation(
-    dry_run: bool = Query(True, description="If true, only report what would happen without executing merges"),
-    db: Session = Depends(get_db)
-):
-    """
-    Run theme consolidation pipeline.
-
-    Steps:
-    1. Update embeddings for all themes
-    2. Find similar pairs via cosine similarity
-    3. Verify with LLM
-    4. Auto-merge high confidence pairs, queue others for review
-
-    Use dry_run=true first to preview changes.
-    """
-    service = ThemeMergingService(db)
-    result = service.run_consolidation(dry_run=dry_run)
-    if not dry_run:
-        from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-        safe_publish_themes_bootstrap_variants()
-    return ConsolidationResultResponse(**result)
-
-
-@router.post("/consolidate/async")
-async def run_theme_consolidation_async(
-    dry_run: bool = Query(False, description="If true, only report what would happen"),
-    db: Session = Depends(get_db)
-):
-    """
-    Run theme consolidation pipeline asynchronously via Celery.
-
-    Returns a task ID for polling status.
-    """
-    from ...tasks.theme_discovery_tasks import consolidate_themes
-
-    task = consolidate_themes.delay(dry_run=dry_run)
-
-    return {
-        "task_id": task.id,
-        "status": "queued",
-        "message": f"Theme consolidation queued (dry_run={dry_run})"
-    }
-
-
-@router.get("/merge-plan/dry-run", response_model=MergePlanDryRunResponse)
-async def get_merge_plan_dry_run(
-    limit_pairs: int = Query(120, ge=1, le=500, description="Max similar pairs to analyze"),
-    pipeline: Optional[str] = Query(None, pattern="^(technical|fundamental)$"),
-    db: Session = Depends(get_db),
-):
-    """
-    Generate a non-mutating duplicate-analysis merge plan with confidence tiers and wave recommendations.
-    """
-    service = ThemeMergingService(db)
-    result = service.generate_dry_run_merge_plan(limit_pairs=limit_pairs, pipeline=pipeline)
-    return MergePlanDryRunResponse(
-        timestamp=result["timestamp"],
-        total_pairs_analyzed=result["total_pairs_analyzed"],
-        confidence_distribution=[MergePlanConfidenceBucketResponse(**row) for row in result["confidence_distribution"]],
-        merge_groups=[MergePlanGroupResponse(**row) for row in result["merge_groups"]],
-        waves=[MergePlanWaveResponse(**row) for row in result["waves"]],
-        ambiguity_clusters=[MergePlanPairResponse(**row) for row in result["ambiguity_clusters"]],
-        do_not_merge=[MergePlanPairResponse(**row) for row in result["do_not_merge"]],
-        manual_review_recommendations=result["manual_review_recommendations"],
-    )
-
-
-@router.post("/embeddings/refresh-campaign", response_model=EmbeddingRefreshCampaignResponse)
-async def run_embedding_refresh_campaign(
-    pipeline: Optional[str] = Query(None, pattern="^(technical|fundamental)$"),
-    refresh_batch_size: int = Query(100, ge=1, le=1000, description="Missing/outdated refresh batch size per pass"),
-    stale_batch_size: int = Query(100, ge=1, le=1000, description="Stale recompute batch size"),
-    stale_max_batches_per_pass: int = Query(10, ge=1, le=100, description="Max stale batches executed per pass"),
-    max_passes: int = Query(6, ge=1, le=50, description="Max bounded passes for campaign"),
-    min_coverage_ratio: float = Query(0.98, ge=0.0, le=1.0, description="Pre-merge coverage gate threshold"),
-    min_freshness_ratio: float = Query(0.95, ge=0.0, le=1.0, description="Pre-merge freshness gate threshold"),
-    db: Session = Depends(get_db),
-):
-    """
-    Execute bounded embedding refresh + stale recompute campaign with gate reporting.
-    """
-    service = ThemeMergingService(db)
-    result = service.run_embedding_refresh_campaign(
-        pipeline=pipeline,
-        refresh_batch_size=refresh_batch_size,
-        stale_batch_size=stale_batch_size,
-        stale_max_batches_per_pass=stale_max_batches_per_pass,
-        max_passes=max_passes,
-        min_coverage_ratio=min_coverage_ratio,
-        min_freshness_ratio=min_freshness_ratio,
-    )
-    return EmbeddingRefreshCampaignResponse(**result)
-
-
-@router.post("/merge-wave/strict-auto", response_model=StrictAutoMergeWaveResponse)
-async def run_strict_auto_merge_wave(
-    pipeline: Optional[str] = Query(None, pattern="^(technical|fundamental)$"),
-    limit_pairs: int = Query(200, ge=1, le=1000, description="Maximum candidate pairs to evaluate"),
-    dry_run: bool = Query(False, description="If true, evaluate strict eligibility without mutating data"),
-    db: Session = Depends(get_db),
-):
-    """
-    Run Wave-1 strict auto-merge for only highest-confidence identical pairs.
-    """
-    service = ThemeMergingService(db)
-    result = service.run_strict_auto_merge_wave(
-        pipeline=pipeline,
-        limit_pairs=limit_pairs,
-        dry_run=dry_run,
-    )
-    if not dry_run:
-        from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-        safe_publish_themes_bootstrap_variants(pipeline)
-    return StrictAutoMergeWaveResponse(**result)
-
-
-@router.post("/merge-wave/manual-review", response_model=ManualReviewWaveResponse)
-async def run_manual_review_wave(
-    payload: ManualReviewWaveRequest,
-    pipeline: Optional[str] = Query(None, pattern="^(technical|fundamental)$"),
-    db: Session = Depends(get_db),
-):
-    """
-    Run Wave-2 moderated manual-review processing with SLA and agreement tracking.
-    """
-    service = ThemeMergingService(db)
-    result = service.run_manual_review_wave(
-        decisions=[row.model_dump() for row in payload.decisions],
-        pipeline=pipeline,
-        sla_target_hours=payload.sla_target_hours,
-        queue_limit=payload.queue_limit,
-        dry_run=payload.dry_run,
-    )
-    if not payload.dry_run:
-        from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-        safe_publish_themes_bootstrap_variants(pipeline)
-    return ManualReviewWaveResponse(**result)
-
-
-@router.get("/candidates/queue", response_model=CandidateThemeQueueResponse)
-async def get_candidate_theme_queue(
-    limit: int = Query(100, ge=1, le=500),
-    offset: int = Query(0, ge=0),
-    pipeline: str = Query("technical", pattern="^(technical|fundamental)$"),
-    db: Session = Depends(get_db),
-):
-    """List candidate themes queued for analyst lifecycle review."""
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-    queue_rows, total_count = service.get_candidate_theme_queue(limit=limit, offset=offset)
-    confidence_bands = [
-        CandidateThemeQueueSummaryBandResponse(**row)
-        for row in service.get_candidate_theme_confidence_bands()
-    ]
-    return CandidateThemeQueueResponse(
-        total=total_count,
-        items=[CandidateThemeQueueItemResponse(**row) for row in queue_rows],
-        confidence_bands=confidence_bands,
-    )
-
-
-@router.post("/candidates/review", response_model=CandidateThemeReviewResponse)
-async def review_candidate_themes(
-    payload: CandidateThemeReviewRequest,
-    pipeline: str = Query("technical", pattern="^(technical|fundamental)$"),
-    db: Session = Depends(get_db),
-):
-    """
-    Bulk analyst triage for candidate themes.
-
-    - `promote` transitions candidate -> active
-    - `reject` transitions candidate -> retired
-    """
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-    result = service.review_candidate_themes(
-        theme_cluster_ids=payload.theme_cluster_ids,
-        action=payload.action,
-        actor=payload.actor or "analyst",
-        note=payload.note,
-    )
-    if not result.get("success"):
-        raise HTTPException(status_code=400, detail=result.get("error", "Candidate review failed"))
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants(pipeline)
-    return CandidateThemeReviewResponse(
-        success=True,
-        action=result["action"],
-        updated=result["updated"],
-        skipped=result["skipped"],
-        results=[CandidateThemeReviewItemResult(**row) for row in result["results"]],
-    )
-
-
-@router.get("/relationship-graph", response_model=ThemeRelationshipGraphResponse)
-async def get_relationship_graph(
-    theme_cluster_id: int = Query(..., ge=1, description="Root theme cluster id"),
-    pipeline: str = Query("technical", pattern="^(technical|fundamental)$"),
-    limit: int = Query(120, ge=1, le=500),
-    db: Session = Depends(get_db),
-):
-    """Return relationship graph nodes/edges centered on the requested theme."""
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_cluster_id).first()
-    if cluster is None:
-        raise HTTPException(status_code=404, detail="Theme not found")
-    if (cluster.pipeline or "technical") != pipeline:
-        raise HTTPException(status_code=404, detail="Theme not found for selected pipeline")
-    service = ThemeDiscoveryService(db, pipeline=pipeline)
-    graph = service.get_theme_relationship_graph(theme_cluster_id, limit=limit)
-    return ThemeRelationshipGraphResponse(
-        theme_cluster_id=theme_cluster_id,
-        total_nodes=len(graph["nodes"]),
-        total_edges=len(graph["edges"]),
-        nodes=[ThemeRelationshipGraphNodeResponse(**node) for node in graph["nodes"]],
-        edges=[ThemeRelationshipGraphEdgeResponse(**edge) for edge in graph["edges"]],
-    )
-
-
-# ==================== Theme Details (parameterized routes - MUST be after static ones) ====================
-
-
-def _safe_rate(numerator: int, denominator: int) -> float:
-    if denominator <= 0:
-        return 0.0
-    return round(float(numerator) / float(denominator), 4)
-
-
-def _confidence_band(value: float | None) -> str:
-    if value is None:
-        return "unknown"
-    if value < 0.40:
-        return "0.00-0.39"
-    if value < 0.70:
-        return "0.40-0.69"
-    if value < 0.85:
-        return "0.70-0.84"
-    return "0.85-1.00"
-
-
-def _score_band(value: float | None) -> str:
-    if value is None:
-        return "unknown"
-    if value < 0.25:
-        return "0.00-0.24"
-    if value < 0.50:
-        return "0.25-0.49"
-    if value < 0.75:
-        return "0.50-0.74"
-    if value < 0.90:
-        return "0.75-0.89"
-    return "0.90-1.00"
-
-
-def _build_band_buckets(
-    rows: list[dict[str, object]],
-    *,
-    band_selector,
-) -> list[ThemeMatchBandBucketResponse]:
-    if not rows:
-        return []
-
-    grouped: dict[str, list[dict[str, object]]] = defaultdict(list)
-    for row in rows:
-        grouped[band_selector(row)].append(row)
-
-    ordered_bands = sorted(grouped.keys())
-    output: list[ThemeMatchBandBucketResponse] = []
-    total = len(rows)
-    for band in ordered_bands:
-        bucket_rows = grouped[band]
-        bucket_total = len(bucket_rows)
-        new_cluster_count = sum(1 for r in bucket_rows if r["match_method"] == "create_new_cluster")
-        attach_count = sum(
-            1
-            for r in bucket_rows
-            if r["match_method"] not in {"create_new_cluster", "unknown"}
-        )
-        output.append(
-            ThemeMatchBandBucketResponse(
-                band=band,
-                count=bucket_total,
-                pct=_safe_rate(bucket_total, total),
-                new_cluster_rate=_safe_rate(new_cluster_count, bucket_total),
-                attach_rate=_safe_rate(attach_count, bucket_total),
+def _fetch_content_items_with_themes_with_recovery(db, **kwargs):
+    """Compatibility wrapper for legacy tests/patches during router decomposition."""
+    try:
+        return _fetch_content_items_with_themes(db, **kwargs)
+    except Exception as exc:
+        if not is_corruption_error(exc):
+            raise
+        safe_rollback(db)
+        _attempt_reindex_theme_content_storage(exc)
+        try:
+            with SessionLocal() as retry_db:
+                return _fetch_content_items_with_themes(retry_db, **kwargs)
+        except Exception as retry_exc:
+            if not is_corruption_error(retry_exc):
+                raise
+            logger.warning(
+                "Theme content browser query still hits corruption after REINDEX; "
+                "checking whether rebuildable theme-content storage can be reset: %s",
+                retry_exc,
             )
-        )
-    return output
-
-
-def _build_match_slice(key: str, rows: list[dict[str, object]]) -> ThemeMatchTelemetrySliceResponse:
-    total = len(rows)
-    new_cluster_count = sum(1 for row in rows if row["match_method"] == "create_new_cluster")
-    attach_count = sum(
-        1
-        for row in rows
-        if row["match_method"] not in {"create_new_cluster", "unknown"}
-    )
-
-    method_counter = Counter(str(row["match_method"] or "unknown") for row in rows)
-    method_distribution = [
-        ThemeMatchMethodDistributionResponse(
-            method=method,
-            count=count,
-            pct=_safe_rate(count, total),
-        )
-        for method, count in sorted(method_counter.items(), key=lambda item: (-item[1], item[0]))
-    ]
-
-    reason_counter = Counter(str(row["match_fallback_reason"] or "none") for row in rows)
-    reason_distribution = [
-        ThemeMatchDecisionReasonDistributionResponse(
-            reason=reason,
-            count=count,
-            pct=_safe_rate(count, total),
-        )
-        for reason, count in sorted(reason_counter.items(), key=lambda item: (-item[1], item[0]))
-    ]
-
-    confidence_bands = _build_band_buckets(rows, band_selector=lambda r: _confidence_band(r["confidence"]))
-    score_bands = _build_band_buckets(rows, band_selector=lambda r: _score_band(r["match_score"]))
-
-    return ThemeMatchTelemetrySliceResponse(
-        key=key,
-        total_mentions=total,
-        new_cluster_count=new_cluster_count,
-        attach_count=attach_count,
-        new_cluster_rate=_safe_rate(new_cluster_count, total),
-        attach_rate=_safe_rate(attach_count, total),
-        method_distribution=method_distribution,
-        decision_reason_distribution=reason_distribution,
-        confidence_bands=confidence_bands,
-        score_bands=score_bands,
-    )
-
-
-@router.get("/matching/telemetry", response_model=ThemeMatchTelemetryResponse)
-async def get_matching_telemetry(
-    days: Annotated[int, Query(ge=1, le=365, description="Rolling window in days")] = 30,
-    pipeline: Annotated[Optional[str], Query(description="Filter by pipeline: technical or fundamental")] = None,
-    source_type: Annotated[Optional[str], Query(description="Filter by source type")] = None,
-    threshold_version: Annotated[Optional[str], Query(description="Filter by threshold version")] = None,
-    db: Session = Depends(get_db),
-):
-    """Aggregate matcher telemetry for ops/model tuning dashboards."""
-    cutoff = datetime.utcnow() - timedelta(days=days)
-
-    query = db.query(
-        ThemeMention.match_method,
-        ThemeMention.match_fallback_reason,
-        ThemeMention.match_score,
-        ThemeMention.confidence,
-        ThemeMention.threshold_version,
-        ThemeMention.source_type,
-        ThemeMention.mentioned_at,
-    ).filter(ThemeMention.mentioned_at >= cutoff)
-
-    if pipeline:
-        query = query.filter(ThemeMention.pipeline == pipeline)
-    if source_type:
-        query = query.filter(ThemeMention.source_type == source_type)
-    if threshold_version:
-        if threshold_version == "unknown":
-            query = query.filter(ThemeMention.threshold_version.is_(None))
-        else:
-            query = query.filter(ThemeMention.threshold_version == threshold_version)
-
-    records = query.all()
-    rows: list[dict[str, object]] = [
-        {
-            "match_method": record.match_method or "unknown",
-            "match_fallback_reason": record.match_fallback_reason,
-            "match_score": record.match_score,
-            "confidence": record.confidence,
-            "threshold_version": record.threshold_version or "unknown",
-            "source_type": record.source_type,
-            "mentioned_at": record.mentioned_at,
-        }
-        for record in records
-    ]
-
-    if not rows:
-        return ThemeMatchTelemetryResponse(
-            window_days=days,
-            start_at=None,
-            end_at=None,
-            pipeline=pipeline,
-            source_type=source_type,
-            threshold_version=threshold_version,
-            total_mentions=0,
-            new_cluster_count=0,
-            attach_count=0,
-            new_cluster_rate=0.0,
-            attach_rate=0.0,
-            method_distribution=[],
-            decision_reason_distribution=[],
-            confidence_bands=[],
-            score_bands=[],
-            by_threshold_version=[],
-            by_source_type=[],
-        )
-
-    overall = _build_match_slice("overall", rows)
-
-    by_threshold: dict[str, list[dict[str, object]]] = defaultdict(list)
-    by_source: dict[str, list[dict[str, object]]] = defaultdict(list)
-    for row in rows:
-        by_threshold[str(row["threshold_version"])].append(row)
-        by_source[str(row["source_type"])].append(row)
-
-    by_threshold_version = [
-        _build_match_slice(key, grouped_rows)
-        for key, grouped_rows in sorted(by_threshold.items(), key=lambda item: (-len(item[1]), item[0]))
-    ]
-    by_source_type = [
-        _build_match_slice(key, grouped_rows)
-        for key, grouped_rows in sorted(by_source.items(), key=lambda item: (-len(item[1]), item[0]))
-    ]
-
-    mentioned_at_values = [row["mentioned_at"] for row in rows if row["mentioned_at"] is not None]
-    start_at = min(mentioned_at_values) if mentioned_at_values else None
-    end_at = max(mentioned_at_values) if mentioned_at_values else None
-
-    return ThemeMatchTelemetryResponse(
-        window_days=days,
-        start_at=start_at,
-        end_at=end_at,
-        pipeline=pipeline,
-        source_type=source_type,
-        threshold_version=threshold_version,
-        total_mentions=overall.total_mentions,
-        new_cluster_count=overall.new_cluster_count,
-        attach_count=overall.attach_count,
-        new_cluster_rate=overall.new_cluster_rate,
-        attach_rate=overall.attach_rate,
-        method_distribution=overall.method_distribution,
-        decision_reason_distribution=overall.decision_reason_distribution,
-        confidence_bands=overall.confidence_bands,
-        score_bands=overall.score_bands,
-        by_threshold_version=by_threshold_version,
-        by_source_type=by_source_type,
-    )
-
-# ==================== L1/L2 Taxonomy Endpoints ====================
-
-@router.get("/taxonomy/l1", response_model=L1ThemeRankingsResponse)
-async def get_l1_rankings(
-    pipeline: str = Query("technical", description="Pipeline filter"),
-    category: Optional[str] = Query(None, description="Filter by L1 category"),
-    sort_by: str = Query("momentum_score", description="Sort field: momentum_score, mentions_7d, num_constituents, basket_return_1w, basket_rs_vs_spy, display_name, rank"),
-    sort_order: str = Query("desc", description="Sort order: asc or desc"),
-    limit: int = Query(100, ge=1, le=500),
-    offset: int = Query(0, ge=0),
-    db: Session = Depends(get_db),
-):
-    """Get L1 theme rankings with aggregated metrics."""
-    from ...services.theme_taxonomy_service import ThemeTaxonomyService
-
-    service = ThemeTaxonomyService(db, pipeline=pipeline)
-    rankings, total = service.get_l1_themes(
-        category_filter=category,
-        limit=limit,
-        offset=offset,
-        sort_by=sort_by,
-        sort_order=sort_order,
-    )
-    return L1ThemeRankingsResponse(
-        total=total,
-        pipeline=pipeline,
-        rankings=[L1ThemeRankingItem(**r) for r in rankings],
-    )
-
-
-@router.get("/taxonomy/l1/{l1_id}/children", response_model=L1ChildrenResponse)
-async def get_l1_children(
-    l1_id: int,
-    sort_by: str = Query("momentum_score", description="Sort field"),
-    sort_order: str = Query("desc", description="Sort order: asc or desc"),
-    limit: int = Query(100, ge=1, le=500),
-    offset: int = Query(0, ge=0),
-    db: Session = Depends(get_db),
-):
-    """Get L2 children of an L1 theme."""
-    from ...services.theme_taxonomy_service import ThemeTaxonomyService
-
-    service = ThemeTaxonomyService(db)
-    result = service.get_l1_with_children(
-        l1_id, limit=limit, offset=offset,
-        sort_by=sort_by, sort_order=sort_order,
-    )
-    if result is None:
-        raise HTTPException(status_code=404, detail=f"L1 theme {l1_id} not found")
-    return L1ChildrenResponse(
-        l1=L1ThemeDetail(**result["l1"]),
-        children=[L1ChildItem(**c) for c in result["children"]],
-        total_children=result["total_children"],
-    )
-
-
-@router.post("/taxonomy/assign")
-async def run_taxonomy_assignment(
-    request: TaxonomyAssignmentRequest,
-    db: Session = Depends(get_db),
-):
-    """Run the full taxonomy assignment pipeline (rules → clustering → LLM naming)."""
-    from ...services.theme_taxonomy_service import ThemeTaxonomyService
-
-    service = ThemeTaxonomyService(db, pipeline=request.pipeline)
-    report = service.run_full_taxonomy_assignment(dry_run=request.dry_run)
-    if not request.dry_run:
-        from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-        safe_publish_themes_bootstrap_variants(request.pipeline)
-    return report
-
-
-@router.post("/taxonomy/assign/async")
-async def run_taxonomy_assignment_async(
-    pipeline: str = Query("technical"),
-    dry_run: bool = Query(False),
-    db: Session = Depends(get_db),
-):
-    """Run taxonomy assignment asynchronously via Celery."""
-    from ...tasks.theme_discovery_tasks import run_taxonomy_assignment as taxonomy_task
-
-    task = taxonomy_task.delay(pipeline=pipeline, dry_run=dry_run)
-    return {"task_id": task.id, "status": "queued"}
-
-
-@router.put("/taxonomy/{l2_id}/reassign")
-async def reassign_l2_to_l1(
-    l2_id: int,
-    request: L2ReassignRequest,
-    db: Session = Depends(get_db),
-):
-    """Manually reassign an L2 theme to a different L1 parent."""
-    from ...services.theme_taxonomy_service import ThemeTaxonomyService
-
-    service = ThemeTaxonomyService(db)
-    success = service.assign_l2_to_l1(l2_id, request.l1_id, method="manual", confidence=1.0)
-    if not success:
-        raise HTTPException(status_code=404, detail="L2 or L1 theme not found")
-    db.commit()
-    from ...services.ui_snapshot_service import safe_publish_themes_bootstrap_variants
-
-    safe_publish_themes_bootstrap_variants()
-    return {"success": True, "l2_id": l2_id, "l1_id": request.l1_id}
-
-
-@router.get("/taxonomy/unassigned", response_model=UnassignedThemesResponse)
-async def get_unassigned_themes(
-    pipeline: str = Query("technical"),
-    limit: int = Query(100, ge=1, le=500),
-    offset: int = Query(0, ge=0),
-    db: Session = Depends(get_db),
-):
-    """Get L2 themes without L1 parent assignment."""
-    from ...services.theme_taxonomy_service import ThemeTaxonomyService
-
-    service = ThemeTaxonomyService(db, pipeline=pipeline)
-    themes, total = service.get_unassigned_themes(limit=limit, offset=offset)
-    return UnassignedThemesResponse(
-        total=total,
-        themes=[UnassignedThemeItem(**t) for t in themes],
-    )
-
-
-@router.get("/taxonomy/categories", response_model=L1CategoriesResponse)
-async def get_l1_categories(
-    pipeline: str = Query("technical"),
-    db: Session = Depends(get_db),
-):
-    """List available L1 categories with theme counts."""
-    from ...services.theme_taxonomy_service import ThemeTaxonomyService
-
-    service = ThemeTaxonomyService(db, pipeline=pipeline)
-    categories = service.get_categories()
-    return L1CategoriesResponse(
-        categories=[L1CategoryItem(**c) for c in categories],
-    )
-
-
-@router.get("/{theme_id}", response_model=ThemeDetailResponse)
-async def get_theme_detail(
-    theme_id: int,
-    db: Session = Depends(get_db)
-):
-    """
-    Get detailed information about a specific theme.
-
-    Includes constituents, metrics, and historical data.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    # Get constituents
-    constituents = db.query(ThemeConstituent).filter(
-        ThemeConstituent.theme_cluster_id == theme_id,
-        ThemeConstituent.is_active == True
-    ).order_by(ThemeConstituent.mention_count.desc()).all()
-
-    # Get latest metrics
-    latest_metrics = db.query(ThemeMetrics).filter(
-        ThemeMetrics.theme_cluster_id == theme_id
-    ).order_by(ThemeMetrics.date.desc()).first()
-
-    service = ThemeDiscoveryService(db, pipeline=cluster.pipeline or "technical")
-    relationships = service.get_theme_relationships(theme_id, limit=50)
-
-    return ThemeDetailResponse(
-        theme=_safe_theme_cluster_response(cluster),
-        constituents=[ThemeConstituentResponse.model_validate(c) for c in constituents],
-        metrics=ThemeMetricsResponse.model_validate(latest_metrics) if latest_metrics else None,
-        relationships=[ThemeRelationshipResponse(**r) for r in relationships],
-    )
-
-
-@router.get("/{theme_id}/history")
-async def get_theme_history(
-    theme_id: int,
-    days: int = Query(30, ge=1, le=180, description="Days of history"),
-    db: Session = Depends(get_db)
-):
-    """
-    Get historical metrics for a theme.
-    """
-    from datetime import timedelta
-
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    cutoff_date = datetime.utcnow().date() - timedelta(days=days)
-
-    metrics = db.query(ThemeMetrics).filter(
-        ThemeMetrics.theme_cluster_id == theme_id,
-        ThemeMetrics.date >= cutoff_date
-    ).order_by(ThemeMetrics.date).all()
-
-    return {
-        "theme": cluster.display_name,
-        "history": [
-            {
-                "date": m.date.isoformat(),
-                "rank": m.rank,
-                "momentum_score": m.momentum_score,
-                "mention_velocity": m.mention_velocity,
-                "basket_rs_vs_spy": m.basket_rs_vs_spy,
-                "status": m.status,
-            }
-            for m in metrics
-        ]
-    }
-
-
-@router.get("/{theme_id}/mentions", response_model=ThemeMentionsResponse)
-async def get_theme_mentions(
-    theme_id: int,
-    limit: int = Query(50, ge=1, le=200, description="Max mentions to return"),
-    db: Session = Depends(get_db)
-):
-    """
-    Get news sources (content items) that mention this theme.
-
-    Returns content items with their associated mention details including
-    excerpt, sentiment, confidence, and source information.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    # Query mentions joined with content items
-    mentions_query = db.query(ThemeMention, ContentItem).join(
-        ContentItem, ThemeMention.content_item_id == ContentItem.id
-    ).filter(
-        ThemeMention.theme_cluster_id == theme_id
-    ).order_by(
-        ThemeMention.mentioned_at.desc()
-    ).limit(limit)
-
-    mentions = mentions_query.all()
-
-    return ThemeMentionsResponse(
-        theme_name=cluster.display_name,
-        theme_id=theme_id,
-        total_count=len(mentions),
-        mentions=[
-            ThemeMentionDetailResponse(
-                mention_id=mention.id,
-                content_title=content.title,
-                content_url=content.url,
-                author=content.author,
-                published_at=content.published_at,
-                excerpt=mention.excerpt,
-                sentiment=mention.sentiment,
-                confidence=mention.confidence,
-                tickers=mention.tickers or [],
-                source_type=mention.source_type,
-                source_name=mention.source_name or content.source_name,
-            )
-            for mention, content in mentions
-        ]
-    )
-
-
-# ==================== Correlation Discovery ====================
-
-@router.get("/correlation/clusters", response_model=CorrelationDiscoveryResponse)
-async def discover_correlation_clusters(
-    correlation_threshold: float = Query(0.6, ge=0.3, le=0.95),
-    min_cluster_size: int = Query(3, ge=2, le=20),
-    db: Session = Depends(get_db)
-):
-    """
-    Discover hidden themes via correlation clustering.
-
-    Finds groups of stocks that move together regardless of industry classification.
-    """
-    service = ThemeCorrelationService(db)
-
-    # Get correlation clusters
-    clusters = service.discover_correlation_clusters(
-        correlation_threshold=correlation_threshold,
-        min_cluster_size=min_cluster_size
-    )
-
-    # Get cross-industry correlations
-    cross_industry = service.find_cross_industry_correlations(
-        min_correlation=correlation_threshold + 0.1
-    )
-
-    return CorrelationDiscoveryResponse(
-        correlation_clusters=[CorrelationClusterResponse(**c) for c in clusters],
-        cross_industry_pairs=[CrossIndustryPairResponse(**p) for p in cross_industry.get("cross_industry_pairs", [])],
-        hub_stocks=cross_industry.get("hub_stocks", [])
-    )
-
-
-@router.get("/{theme_id}/validate", response_model=ThemeValidationResponse)
-async def validate_theme(
-    theme_id: int,
-    min_correlation: float = Query(0.5, ge=0.2, le=0.9),
-    db: Session = Depends(get_db)
-):
-    """
-    Validate a theme by checking internal correlations.
-
-    A valid theme has high average correlation between constituents.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    service = ThemeCorrelationService(db)
-    result = service.validate_theme(theme_id, min_correlation)
-
-    return ThemeValidationResponse(
-        theme=cluster.display_name,
-        **result
-    )
-
-
-@router.get("/{theme_id}/entrants")
-async def find_theme_entrants(
-    theme_id: int,
-    correlation_threshold: float = Query(0.6, ge=0.4, le=0.9),
-    db: Session = Depends(get_db)
-):
-    """
-    Find stocks that may be joining this theme.
-
-    Identifies stocks with high correlation to the theme basket
-    that aren't yet constituents.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    service = ThemeCorrelationService(db)
-    entrants = service.find_new_theme_entrants(
-        theme_id,
-        correlation_threshold=correlation_threshold
-    )
-
-    return {
-        "theme": cluster.display_name,
-        "potential_entrants": entrants
-    }
-
-
-@router.get("/{theme_id}/similar", response_model=SimilarThemesResponse)
-async def find_similar_themes(
-    theme_id: int,
-    threshold: float = Query(0.75, ge=0.5, le=0.99, description="Minimum similarity threshold"),
-    db: Session = Depends(get_db)
-):
-    """
-    Find themes similar to the given theme using embedding similarity.
-
-    Uses sentence-transformers embeddings and cosine similarity.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    service = ThemeMergingService(db)
-    similar = service.find_similar_themes(theme_id, threshold=threshold)
-
-    return SimilarThemesResponse(
-        source_theme_id=theme_id,
-        source_theme_name=cluster.display_name,
-        threshold=threshold,
-        similar_themes=[SimilarThemeResponse(**s) for s in similar]
-    )
-
-
-# ==================== Alert Operations ====================
-
-@router.post("/alerts/check")
-async def check_for_alerts(
-    db: Session = Depends(get_db)
-):
-    """
-    Check for and generate new alerts.
-    """
-    service = ThemeDiscoveryService(db)
-    alerts = service.check_for_alerts()
-
-    return {
-        "new_alerts": len(alerts),
-        "alerts": [{"type": a.alert_type, "title": a.title} for a in alerts]
-    }
-
-
-@router.post("/alerts/{alert_id}/read")
-async def mark_alert_read(
-    alert_id: int,
-    db: Session = Depends(get_db)
-):
-    """
-    Mark an alert as read.
-    """
-    alert = db.query(ThemeAlert).filter(ThemeAlert.id == alert_id).first()
-    if not alert:
-        raise HTTPException(status_code=404, detail="Alert not found")
-
-    alert.is_read = True
-    alert.read_at = datetime.utcnow()
-    db.commit()
-
-    return {"status": "marked as read"}
-
-
-# ==================== Theme Management ====================
-
-@router.post("/create-from-cluster")
-async def create_theme_from_cluster(
-    name: str,
-    symbols: list[str],
-    description: Optional[str] = None,
-    db: Session = Depends(get_db)
-):
-    """
-    Create a new theme from a correlation cluster.
-    """
-    service = ThemeCorrelationService(db)
-    cluster = service.create_theme_from_cluster(
-        symbols=symbols,
-        name=name,
-        description=description
-    )
-
-    return {
-        "status": "created",
-        "theme_id": cluster.id,
-        "name": cluster.display_name,
-        "num_constituents": len(symbols)
-    }
-
-
-@router.post("/{theme_id}/add-constituents")
-async def add_theme_constituents(
-    theme_id: int,
-    symbols: list[str],
-    db: Session = Depends(get_db)
-):
-    """
-    Add stocks to a theme manually.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    added = 0
-    for symbol in symbols:
-        existing = db.query(ThemeConstituent).filter(
-            ThemeConstituent.theme_cluster_id == theme_id,
-            ThemeConstituent.symbol == symbol
-        ).first()
-
-        if not existing:
-            constituent = ThemeConstituent(
-                theme_cluster_id=theme_id,
-                symbol=symbol,
-                source="manual",
-                confidence=1.0,
-                mention_count=0,
-                first_mentioned_at=datetime.utcnow(),
-                last_mentioned_at=datetime.utcnow(),
-            )
-            db.add(constituent)
-            added += 1
-
-    db.commit()
-
-    return {"status": "success", "added": added}
-
-
-@router.delete("/{theme_id}")
-async def deactivate_theme(
-    theme_id: int,
-    db: Session = Depends(get_db)
-):
-    """
-    Deactivate a theme.
-    """
-    cluster = db.query(ThemeCluster).filter(ThemeCluster.id == theme_id).first()
-    if not cluster:
-        raise HTTPException(status_code=404, detail="Theme not found")
-
-    cluster.is_active = False
-    db.commit()
-
-    return {"status": "deactivated", "theme": cluster.display_name}
+            if not _corruption_targets_theme_content_storage(**kwargs):
+                raise
+            _reset_corrupt_theme_content_storage(retry_exc)
+            with SessionLocal() as reset_retry_db:
+                return _fetch_content_items_with_themes(reset_retry_db, **kwargs)
+
+# Keep static/source/pipeline paths before parameterized theme routes.
+router.include_router(content_sources_router)
+router.include_router(content_pipeline_router)
+router.include_router(review_merge_router)
+router.include_router(taxonomy_router)
+router.include_router(queries_router)
