+++ backend/app/theme_platform/__init__.py
@@ -1,0 +1,2 @@
+"""Theme platform application slice."""
+
