+++ backend/tests/unit/test_theme_twitter_session_api.py
@@ -7,11 +7,13 @@ import pytest
 import pytest_asyncio
 
 from app.main import app
+from app.services import server_auth
 from app.services.xui_session_bridge_service import XUISessionBridgeError
 
 
 @pytest_asyncio.fixture
-async def client():
+async def client(monkeypatch: pytest.MonkeyPatch):
+    monkeypatch.setattr(server_auth.settings, "server_auth_enabled", False)
     transport = httpx.ASGITransport(app=app)
     async with httpx.AsyncClient(transport=transport, base_url="http://test") as c:
         yield c
